--- a/test/Microsoft.TestPlatform.PerformanceTests/TranslationLayer/DiscoveryPerfTests.cs
+++ b/test/Microsoft.TestPlatform.PerformanceTests/TranslationLayer/DiscoveryPerfTests.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Collections.Generic;
+
 using Microsoft.TestPlatform.PerformanceTests.PerfInstrumentation;
 using Microsoft.VisualStudio.TestPlatform.Common.Telemetry;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Client;
@@ -25,16 +27,22 @@ public class DiscoveryPerfTests : TelemetryPerfTestBase
     [DataRow("XUnit100Passing", 100)]
     [DataRow("XUnit1000Passing", 1000)]
     [DataRow("XUnit10kPassing", 10_000)]
+    [DataRow("Perfy.TestAdapter", 1)]
+    [DataRow("Perfy.TestAdapter", 100)]
+    [DataRow("Perfy.TestAdapter", 1000)]
+    [DataRow("Perfy.TestAdapter", 10_000)]
     public void DiscoverTests(string projectName, double expectedNumberOfTests)
     {
         var discoveryEventHandler2 = new DiscoveryEventHandler2();
-        TestPlatformOptions options = new() { CollectMetrics = true };
-
+        TestPlatformOptions options = new() { CollectMetrics = true, };
         var perfAnalyzer = new PerfAnalyzer();
         using (perfAnalyzer.Start())
         {
-            var vstestConsoleWrapper = GetVsTestConsoleWrapper();
-            vstestConsoleWrapper.DiscoverTests(GetPerfAssetFullPath(projectName), GetDefaultRunSettings(), options, discoveryEventHandler2);
+            // This tells to PerfyTestAdapter how many tests it should return, this is our overhead baseline.
+            var perfyTestAdapterEnv = new Dictionary<string, string?> { ["TEST_COUNT"] = expectedNumberOfTests.ToString() };
+            var vstestConsoleWrapper = GetVsTestConsoleWrapper(perfyTestAdapterEnv, traceLevel: System.Diagnostics.TraceLevel.Off);
+            var assetPath = GetPerfAssetFullPath(projectName);
+            vstestConsoleWrapper.DiscoverTests(assetPath, GetDefaultRunSettings(), options, discoveryEventHandler2);
             vstestConsoleWrapper.EndSession();
         }
         Assert.AreEqual(expectedNumberOfTests, discoveryEventHandler2.Metrics[TelemetryDataConstants.TotalTestsDiscovered]);
@@ -56,6 +64,10 @@ public class DiscoveryPerfTests : TelemetryPerfTestBase
     [DataRow("XUnit100Passing", 100)]
     [DataRow("XUnit1000Passing", 1000)]
     [DataRow("XUnit10kPassing", 10_000)]
+    [DataRow("Perfy.TestAdapter", 1)]
+    [DataRow("Perfy.TestAdapter", 100)]
+    [DataRow("Perfy.TestAdapter", 1000)]
+    [DataRow("Perfy.TestAdapter", 10_000)]
     public void DiscoverTestsWithDefaultAdaptersSkipped(string projectName, double expectedNumberOfTests)
     {
 
@@ -69,7 +81,9 @@ public class DiscoveryPerfTests : TelemetryPerfTestBase
         var perfAnalyzer = new PerfAnalyzer();
         using (perfAnalyzer.Start())
         {
-            var vstestConsoleWrapper = GetVsTestConsoleWrapper();
+            // This tells to PerfyTestAdapter how many tests it should return, this is our overhead baseline.
+            var perfyTestAdapterEnv = new Dictionary<string, string?> { ["TEST_COUNT"] = expectedNumberOfTests.ToString() };
+            var vstestConsoleWrapper = GetVsTestConsoleWrapper(perfyTestAdapterEnv, traceLevel: System.Diagnostics.TraceLevel.Off);
             vstestConsoleWrapper.DiscoverTests(GetPerfAssetFullPath(projectName), GetDefaultRunSettings(), options, discoveryEventHandler2);
             vstestConsoleWrapper.EndSession();
         }