--- a/test/Microsoft.TestPlatform.PerformanceTests/TranslationLayer/ExecutionPerfTests.cs
+++ b/test/Microsoft.TestPlatform.PerformanceTests/TranslationLayer/ExecutionPerfTests.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Collections.Generic;
+
 using Microsoft.TestPlatform.PerformanceTests.PerfInstrumentation;
 using Microsoft.VisualStudio.TestPlatform.Common.Telemetry;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Client;
@@ -25,6 +27,10 @@ public class ExecutionPerfTests : TelemetryPerfTestBase
     [DataRow("XUnit100Passing", 100)]
     [DataRow("XUnit1000Passing", 1000)]
     [DataRow("XUnit10kPassing", 10_000)]
+    [DataRow("Perfy.TestAdapter", 1)]
+    [DataRow("Perfy.TestAdapter", 100)]
+    [DataRow("Perfy.TestAdapter", 1000)]
+    [DataRow("Perfy.TestAdapter", 10_000)]
     public void RunTests(string projectName, double expectedNumberOfTests)
     {
         var runEventHandler = new RunEventHandler();
@@ -33,8 +39,9 @@ public class ExecutionPerfTests : TelemetryPerfTestBase
         var perfAnalyzer = new PerfAnalyzer();
         using (perfAnalyzer.Start())
         {
-            var vstestConsoleWrapper = GetVsTestConsoleWrapper(traceLevel: System.Diagnostics.TraceLevel.Off);
-
+            // This tells to PerfyTestAdapter how many tests it should return, this is our overhead baseline.
+            var perfyTestAdapterEnv = new Dictionary<string, string?> { ["TEST_COUNT"] = expectedNumberOfTests.ToString() };
+            var vstestConsoleWrapper = GetVsTestConsoleWrapper(perfyTestAdapterEnv, traceLevel: System.Diagnostics.TraceLevel.Off);
             vstestConsoleWrapper.RunTests(GetPerfAssetFullPath(projectName), GetDefaultRunSettings(), options, runEventHandler);
             vstestConsoleWrapper.EndSession();
         }
@@ -57,6 +64,10 @@ public class ExecutionPerfTests : TelemetryPerfTestBase
     [DataRow("XUnit100Passing", 100)]
     [DataRow("XUnit1000Passing", 1000)]
     [DataRow("XUnit10kPassing", 10_000)]
+    [DataRow("Perfy.TestAdapter", 1)]
+    [DataRow("Perfy.TestAdapter", 100)]
+    [DataRow("Perfy.TestAdapter", 1000)]
+    [DataRow("Perfy.TestAdapter", 10_000)]
     public void RunTestsWithDefaultAdaptersSkipped(string projectName, double expectedNumberOfTests)
     {
         var runEventHandler = new RunEventHandler();
@@ -69,7 +80,9 @@ public class ExecutionPerfTests : TelemetryPerfTestBase
         var perfAnalyzer = new PerfAnalyzer();
         using (perfAnalyzer.Start())
         {
-            var vstestConsoleWrapper = GetVsTestConsoleWrapper(traceLevel: System.Diagnostics.TraceLevel.Off);
+            // This tells to PerfyTestAdapter how many tests it should return, this is our overhead baseline.
+            var perfyTestAdapterEnv = new Dictionary<string, string?> { ["TEST_COUNT"] = expectedNumberOfTests.ToString() };
+            var vstestConsoleWrapper = GetVsTestConsoleWrapper(perfyTestAdapterEnv, traceLevel: System.Diagnostics.TraceLevel.Off);
             vstestConsoleWrapper.RunTests(GetPerfAssetFullPath(projectName), GetDefaultRunSettings(), options, runEventHandler);
             vstestConsoleWrapper.EndSession();
         }