--- a/test/TestAssets/performance/Perfy.TestAdapter/Perfy.TestAdapter.csproj
+++ b/test/TestAssets/performance/Perfy.TestAdapter/Perfy.TestAdapter.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <!-- Package dependency versions -->
+  <Import Project="..\..\..\..\scripts\build\TestAssets.props" />
+  <PropertyGroup>
+    <TargetFrameworks>net6.0;net48</TargetFrameworks>
+    <TargetFrameworks Condition=" '$(DotNetBuildFromSource)' == 'true' ">netcoreapp3.1</TargetFrameworks>
+    <CheckEolTargetFramework>false</CheckEolTargetFramework>
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(NETTestSdkVersion)" />
+    <!-- Don't output any of this into the output folder,
+    it would clash with Test.Sdk, MSTest also does not ship ObjectModel that it builds against. -->
+    <PackageReference Include="Microsoft.TestPlatform.ObjectModel" Version="$(NETTestSdkVersion)" PrivateAssets="All" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
+  </ItemGroup>
+
+</Project>