… test/TestAssets/TestAssets.sln