## Description
Adds test adapter called Perfy that has almost no overhead to the perf tests. This will help us separate TP and adapter overhead. TEST_COUNT env variable is set for the adapter which generates that many test cases, or test results. 

This is work is common to all adapters, and is reasonably cheap, any overhead that is measured in the run, can be attributed to TestPlatform. In our results we can then see how much time it took for Perfy to run 1, 100, 1000, and 10k tests, and compare that with other adapters, which will show e.g. 1.1s for Perfy and 1.5s for NUnit, meaning that 1.1s is TestPlatform overhead, and 400ms is NUnit overhead.